子コンポーネント化
