+++ app/components/form/index.ts
@@ -1,0 +1,2 @@
+export { default as Input } from './_input';
+export { default as ErrorText } from './_errorText';
