+++ app/features/main/home/homeNest/child00/_component.tsx
@@ -1,72 +1,8 @@
 import { type FC, useCallback, useMemo, useRef, useState } from 'react';
-import { type FieldValues, useController } from 'react-hook-form';
-import { Keyboard, Pressable, StyleSheet, Text, TextInput, View } from 'react-native';
+import { Keyboard, Pressable, StyleSheet, Text, View } from 'react-native';
 import RNPickerSelect from 'react-native-picker-select';
 
-import type {
-  TypeErrorText,
-  TypeInputProps,
-  TypePickerField,
-  TypePickerSelectStyles,
-  TypeResultArea,
-} from './_type';
-
-/* -----------------------------------------------
- * テキストインプット
- * ----------------------------------------------- */
-export const Input = <TFieldValues extends FieldValues>({
-  containerStyle,
-  control,
-  errorText,
-  label,
-  name,
-  rules,
-  style,
-  ...textInputProps
-}: TypeInputProps<TFieldValues>) => {
-  const {
-    field: { onBlur, onChange, value },
-  } = useController({ control, name, rules });
-
-  const inputValue = typeof value === 'string' ? value : '';
-  const hasError = errorText?.message != null;
-
-  return (
-    <View style={[inputStyles.container, containerStyle]}> 
-      <Text style={inputStyles.label}>{label}</Text>
-      <TextInput
-        {...textInputProps}
-        onBlur={onBlur}
-        onChangeText={onChange}
-        style={[inputStyles.input, hasError ? inputStyles.inputError : null, style]}
-        value={inputValue}
-      />
-      <ErrorText {...errorText} />
-    </View>
-  );
-};
-
-const inputStyles = StyleSheet.create({
-  container: {
-    marginBottom: 16,
-  },
-  label: {
-    fontSize: 14,
-    fontWeight: '500',
-    marginBottom: 8,
-  },
-  input: {
-    backgroundColor: '#fff',
-    borderColor: '#d6d6d6',
-    borderRadius: 8,
-    borderWidth: 1,
-    paddingHorizontal: 12,
-    paddingVertical: 10,
-  },
-  inputError: {
-    borderColor: '#e53935',
-  },
-});
+import type { TypePickerField, TypePickerSelectStyles, TypeResultArea } from './_type';
 
 /* -----------------------------------------------
  * セレクトボックス
@@ -197,23 +133,6 @@ const pickerSelectStyles: TypePickerSelectStyles = {
     fontWeight: '600',
   },
 };
-
-/* -----------------------------------------------
- * エラーテキスト
- * ----------------------------------------------- */
-export const ErrorText: React.FC<TypeErrorText> = (errorsType) => {
-  if (errorsType.message == null) {
-    return;
-  }
-  return <Text style={errorStyles.text}>{errorsType.message}</Text>;
-};
-const errorStyles = StyleSheet.create({
-  text: {
-    color: '#e53935',
-    fontSize: 12,
-    marginTop: 4,
-  },
-});
 
 /* -----------------------------------------------
  * submit 出力結果表示エリア
