+++ app/lib/types/typeComponents/_form.ts
@@ -1,0 +1,27 @@
+import type {
+  Control,
+  FieldError,
+  FieldValues,
+  Merge,
+  Path,
+  RegisterOptions,
+} from 'react-hook-form';
+import type { StyleProp, TextInputProps, TextStyle, ViewStyle } from 'react-native';
+
+/* -----------------------------------------------
+ * type エラーテキスト
+ * ----------------------------------------------- */
+export type TypeErrorText = Merge<FieldError, (FieldError | undefined)[]>;
+
+/* -----------------------------------------------
+ * type インプット
+ * ----------------------------------------------- */
+export type TypeInputProps<TFieldValues extends FieldValues> = {
+  label: string;
+  control: Control<TFieldValues>;
+  name: Path<TFieldValues>;
+  rules?: RegisterOptions<TFieldValues, Path<TFieldValues>>;
+  errorText?: TypeErrorText | FieldError;
+  containerStyle?: StyleProp<ViewStyle>;
+  style?: StyleProp<TextStyle>;
+} & Omit<TextInputProps, 'onBlur' | 'onChangeText' | 'value' | 'style'>;
