+++ app/lib/types/typeComponents/index.ts
@@ -1,2 +1,3 @@
+export * from './_form';
 export * from './_layout';
 export * from './_svg';
